feat: add dynamic [post] route
