+++ components/PostPreviews/types.ts
@@ -1,4 +1,5 @@
 export interface PostTypes {
+  id: number
   _id?: string
   date?: Date
   user_id?: number
